requirements.txt
openai-whisper>=20231117
ffmpeg-python>=0.2.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
deep-translator>=1.11.4
numpy>=1.24.0
psutil>=5.9.0
plyer>=2.1.0
